fix: moved Enum classes from main.py to model.py.
- main fix: moved Enum classes
- read_item() was misnamed due to copy/paste
